+++ src/devices/Common/Iot/Device/Graphics/BitmapImage.cs
@@ -1,6 +1,11 @@
-﻿using System;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
 
-namespace Iot.Device.Bindings.Utils
+using System;
+using System.Drawing;
+
+namespace Iot.Device.Graphics
 {
     public abstract class BitmapImage
     {
