+++ src/devices/Ws2812B/samples/rpi-neo-pixels_bb.png
